Create README.md
Base64.net is a WindowsForms application written on C# programming language. The purpose  of the app is encoding/decoding selected file using Base64.
The algorithm of Base64 encoding/decoding is not optimised. There are numbers of ways to make such work more flexible (see Wiki for example). As for this case, this program uses algorithms which i made up by my own during solving this computer problem